+++ vortex-array/src/array/chunked/compute/mod.rs
@@ -73,7 +73,7 @@ impl CastFn<ChunkedArray> for ChunkedEncoding {
 
 #[cfg(test)]
 mod test {
-    use vortex_buffer::ScalarBufferMut;
+    use vortex_buffer::BufferMut;
     use vortex_dtype::{DType, Nullability, PType};
 
     use crate::array::chunked::ChunkedArray;
@@ -111,7 +111,7 @@ mod test {
             .into_primitive()
             .unwrap()
             .into_maybe_null_slice::<u64>(),
-            ScalarBufferMut::<u64>::from_iter([0u64, 1, 2, 3]),
+            BufferMut::<u64>::from_iter([0u64, 1, 2, 3]),
         );
     }
 }