+++ vortex-array/src/array/constant/canonical.rs
@@ -1,6 +1,6 @@
 use arrow_array::builder::make_view;
 use arrow_buffer::BooleanBuffer;
-use vortex_buffer::{ByteBufferMut, ScalarBuffer};
+use vortex_buffer::{Buffer, ByteBufferMut};
 use vortex_dtype::{match_each_native_ptype, DType, Nullability};
 use vortex_error::{vortex_bail, VortexResult};
 use vortex_scalar::{BinaryScalar, BoolScalar, ExtScalar, Utf8Scalar};
@@ -90,7 +90,7 @@ fn canonical_byte_view(
             if scalar_bytes.len() >= BinaryView::MAX_INLINED_SIZE {
                 buffers.push(
                     PrimitiveArray::new(
-                        ScalarBuffer::copy_from_vec(scalar_bytes.to_vec()),
+                        Buffer::copy_from_vec(scalar_bytes.to_vec()),
                         Validity::NonNullable,
                     )
                     .into_array(),
@@ -107,11 +107,9 @@ fn canonical_byte_view(
             for _ in 0..len {
                 views.extend_from_slice(&view.to_le_bytes());
             }
-            let views = PrimitiveArray::new(
-                ScalarBuffer::<u8>::from(views.freeze()),
-                Validity::NonNullable,
-            )
-            .into_array();
+            let views =
+                PrimitiveArray::new(Buffer::<u8>::from(views.freeze()), Validity::NonNullable)
+                    .into_array();
 
             let validity = if dtype.nullability() == Nullability::NonNullable {
                 Validity::NonNullable