+++ vortex-array/src/array/varbinview/mod.rs
@@ -6,7 +6,7 @@ use ::serde::{Deserialize, Serialize};
 use arrow_array::builder::{BinaryViewBuilder, GenericByteViewBuilder, StringViewBuilder};
 use arrow_array::types::{BinaryViewType, ByteViewType, StringViewType};
 use arrow_array::{ArrayRef, BinaryViewArray, GenericByteViewArray, StringViewArray};
-use arrow_buffer::ScalarBuffer;
+use arrow_buffer::Buffer;
 use itertools::Itertools;
 use static_assertions::{assert_eq_align, assert_eq_size};
 use vortex_buffer::ByteBuffer;
@@ -574,14 +574,14 @@ pub(crate) fn varbinview_as_arrow(var_bin_view: &VarBinViewArray) -> ArrayRef {
     match var_bin_view.dtype() {
         DType::Binary(_) => Arc::new(unsafe {
             BinaryViewArray::new_unchecked(
-                ScalarBuffer::<u128>::from(views.buffer().clone().into_arrow_buffer()),
+                Buffer::<u128>::from(views.buffer().clone().into_arrow_buffer()),
                 data,
                 nulls,
             )
         }),
         DType::Utf8(_) => Arc::new(unsafe {
             StringViewArray::new_unchecked(
-                ScalarBuffer::<u128>::from(views.buffer().clone().into_arrow_buffer()),
+                Buffer::<u128>::from(views.buffer().clone().into_arrow_buffer()),
                 data,
                 nulls,
             )