+++ vortex-array/src/arrow/array.rs
@@ -14,7 +14,7 @@ use arrow_array::types::{
 };
 use arrow_array::{BinaryViewArray, GenericByteViewArray, GenericListArray, StringViewArray};
 use arrow_buffer::buffer::{NullBuffer, OffsetBuffer};
-use arrow_buffer::{ArrowNativeType, BooleanBuffer, Buffer, ScalarBuffer};
+use arrow_buffer::{ArrowNativeType, BooleanBuffer, Buffer, Buffer};
 use arrow_schema::{DataType, TimeUnit as ArrowTimeUnit};
 use itertools::Itertools;
 use vortex_buffer::{Alignment, ByteBuffer};
@@ -48,13 +48,13 @@ impl From<BooleanBuffer> for ArrayData {
     }
 }
 
-impl<T> From<ScalarBuffer<T>> for ArrayData
+impl<T> From<Buffer<T>> for ArrayData
 where
     T: ArrowNativeType + NativePType,
 {
-    fn from(value: ScalarBuffer<T>) -> Self {
+    fn from(value: Buffer<T>) -> Self {
         PrimitiveArray::new(
-            vortex_buffer::ScalarBuffer::<T>::from_arrow(value, Alignment::of::<T>()),
+            vortex_buffer::Buffer::<T>::from_arrow(value, Alignment::of::<T>()),
             Validity::NonNullable,
         )
         .into_array()
@@ -67,7 +67,7 @@ where
 {
     fn from(value: OffsetBuffer<O>) -> Self {
         let primitive = PrimitiveArray::new(
-            vortex_buffer::ScalarBuffer::from_arrow(value.into_inner(), Alignment::of::<O>()),
+            vortex_buffer::Buffer::from_arrow(value.into_inner(), Alignment::of::<O>()),
             Validity::NonNullable,
         );
         primitive.statistics().set(Stat::IsSorted, true.into());
@@ -84,7 +84,7 @@ where
 {
     fn from_arrow(value: &ArrowPrimitiveArray<T>, nullable: bool) -> Self {
         let arr = PrimitiveArray::new(
-            vortex_buffer::ScalarBuffer::from_arrow(value.values().clone(), Alignment::of::<T>()),
+            vortex_buffer::Buffer::from_arrow(value.values().clone(), Alignment::of::<T>()),
             nulls(value.nulls(), nullable),
         );