+++ vortex-buffer/benches/vortex_buffer.rs
@@ -3,25 +3,25 @@
 use std::iter::Iterator;
 
 use arrow_buffer::ArrowNativeType;
-use vortex_buffer::ScalarBuffer;
+use vortex_buffer::Buffer;
 
 fn main() {
     divan::main();
 }
 
-// We wrap the Arrow ScalarBuffer so divan output distinguishes the type name.
-pub struct ArrowScalarBuffer<T: ArrowNativeType>(pub arrow_buffer::ScalarBuffer<T>);
+// We wrap the Arrow Buffer so divan output distinguishes the type name.
+pub struct ArrowBuffer<T: ArrowNativeType>(pub arrow_buffer::Buffer<T>);
 
-impl<T: ArrowNativeType> FromIterator<T> for ArrowScalarBuffer<T> {
+impl<T: ArrowNativeType> FromIterator<T> for ArrowBuffer<T> {
     fn from_iter<I: IntoIterator<Item = T>>(iter: I) -> Self {
-        Self(arrow_buffer::ScalarBuffer::from_iter(iter))
+        Self(arrow_buffer::Buffer::from_iter(iter))
     }
 }
 
 #[divan::bench(
     types = [
-        ArrowScalarBuffer<i32>,
-        ScalarBuffer<i32>,
+        ArrowBuffer<i32>,
+        Buffer<i32>,
     ],
     args = [1, 100, 10_00, 100_000, 10_000_000],
 )]