+++ vortex-buffer/src/arrow.rs
@@ -3,17 +3,17 @@ use bytes::Bytes;
 use vortex_dtype::NativePType;
 use vortex_error::vortex_panic;
 
-use crate::{Alignment, ByteBuffer, ScalarBuffer};
+use crate::{Alignment, Buffer, ByteBuffer};
 
-impl<T: NativePType + ArrowNativeType> ScalarBuffer<T> {
-    /// Converts the buffer zero-copy into a `arrow_buffer::ScalarBuffer`.
-    pub fn into_arrow(self) -> arrow_buffer::ScalarBuffer<T> {
+impl<T: NativePType + ArrowNativeType> Buffer<T> {
+    /// Converts the buffer zero-copy into a `arrow_buffer::Buffer`.
+    pub fn into_arrow(self) -> arrow_buffer::Buffer<T> {
         let bytes = self.into_inner();
         // This is cheeky. But it uses From<bytes::Bytes> for arrow_buffer::Bytes, even though
         // arrow_buffer::Bytes is only pub(crate). Seems weird...
         // See: https://github.com/apache/arrow-rs/issues/6033
         let buffer = arrow_buffer::Buffer::from_bytes(bytes.into());
-        arrow_buffer::ScalarBuffer::from(buffer)
+        arrow_buffer::Buffer::from(buffer)
     }
 
     /// Convert an Arrow scalar buffer into a Vortex scalar buffer.
@@ -21,7 +21,7 @@ impl<T: NativePType + ArrowNativeType> ScalarBuffer<T> {
     /// ## Panics
     ///
     /// Panics if the Arrow buffer is not sufficiently aligned.
-    pub fn from_arrow(arrow: arrow_buffer::ScalarBuffer<T>, alignment: Alignment) -> Self {
+    pub fn from_arrow(arrow: arrow_buffer::Buffer<T>, alignment: Alignment) -> Self {
         let length = arrow.len();
 
         let bytes = Bytes::from_owner(ArrowWrapper(arrow.into_inner()));