+++ vortex-buffer/src/const.rs
@@ -2,23 +2,23 @@ use std::ops::Deref;
 
 use vortex_error::{vortex_bail, VortexError};
 
-use crate::{Alignment, ScalarBuffer};
+use crate::{Alignment, Buffer};
 
 /// An aligned buffer with compile-time alignment.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd)]
-pub struct ConstBuffer<T, const A: usize>(ScalarBuffer<T>);
+pub struct ConstBuffer<T, const A: usize>(Buffer<T>);
 
 impl<T, const A: usize> ConstBuffer<T, A> {
     /// Unwrap the inner buffer.
-    pub fn into_inner(self) -> ScalarBuffer<T> {
+    pub fn into_inner(self) -> Buffer<T> {
         self.0
     }
 }
 
-impl<T, const A: usize> TryFrom<ScalarBuffer<T>> for ConstBuffer<T, A> {
+impl<T, const A: usize> TryFrom<Buffer<T>> for ConstBuffer<T, A> {
     type Error = VortexError;
 
-    fn try_from(value: ScalarBuffer<T>) -> Result<Self, Self::Error> {
+    fn try_from(value: Buffer<T>) -> Result<Self, Self::Error> {
         if !value.alignment().is_aligned_to(Alignment::new(A)) {
             vortex_bail!(
                 "Cannot convert buffer with alignment {} to buffer with alignment {}",
@@ -30,8 +30,8 @@ impl<T, const A: usize> TryFrom<ScalarBuffer<T>> for ConstBuffer<T, A> {
     }
 }
 
-impl<T, const A: usize> AsRef<ScalarBuffer<T>> for ConstBuffer<T, A> {
-    fn as_ref(&self) -> &ScalarBuffer<T> {
+impl<T, const A: usize> AsRef<Buffer<T>> for ConstBuffer<T, A> {
+    fn as_ref(&self) -> &Buffer<T> {
         &self.0
     }
 }